requirements.txt
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==10.0.0
rembg==2.0.50
Flask==2.0.1
shopify==9.0.0
python-dotenv==0.19.0
requests==2.26.0
Werkzeug==2.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.1 